Opt: faster gen m3u8

+++ protocol/hls/item.go
@@ -2,12 +2,12 @@ package hls
 
 type TSItem struct {
 	TsName   string
-	SeqNum   int
-	Duration int
+	SeqNum   int64
+	Duration int64
 	Data     []byte
 }
 
-func NewTSItem(tsName string, duration, seqNum int, b []byte) *TSItem {
+func NewTSItem(tsName string, duration, seqNum int64, b []byte) *TSItem {
 	var item = new(TSItem)
 	item.TsName = tsName
 	item.SeqNum = seqNum
